requirements.txt
torch>=1.13.0
torchvision
transformers
diffusers
opencv-python
matplotlib
Pillow
numpy